DeeplabV3-based method/requirements.txt
colorama==0.4.6
matplotlib==3.6.2
numpy==1.26.4
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
pandas==2.1.3
Pillow==9.3.0
Pillow==10.3.0
scikit_learn==1.4.1.post1
torch==2.2.1+cu121
torchvision==0.16.1+cu121
tqdm==4.66.2